Move legacy pprinter to next.js
