--- packages/liphe/src/lib.d.ts
@@ -1,4 +1,0 @@
-declare module "raw-loader!delisp/lib/prelude.dl" {
-  const content: string;
-  export = content;
-}
